Move transcribe/translate queries to workflow api
