+++ source/workflowapi/external_resources.json
@@ -167,6 +167,20 @@
                   ]
                 },
                 {
+                  "Effect": "Allow",
+                  "Action": [
+                    "transcribe:CreateVocabulary",
+                    "transcribe:DeleteVocabulary",
+                    "transcribe:GetVocabulary",
+                    "transcribe:ListVocabularies",
+                    "translate:DeleteTerminology",
+                    "translate:GetTerminology",
+                    "translate:ImportTerminology",
+                    "translate:ListTerminologies"
+                  ],
+                  "Resource": "*"
+                },
+                {
                   "Action": [
                     "logs:CreateLogGroup",
                     "logs:CreateLogStream",
